+++ lib/src/local_working_copy.rs
@@ -1150,7 +1150,7 @@ impl TreeState {
                 return Err(SnapshotError::Other {
                     message: "Failed to query the filesystem monitor".to_string(),
                     err: "Cannot query Watchman because jj was not compiled with the `watchman` \
-                          feature (consider disabling `core.fsmonitor`)"
+                          feature (consider disabling `fsmonitor.backend`)"
                         .into(),
                 });
             }