+++ CHANGELOG.md
@@ -33,6 +33,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   planned to be removed in a future release.
   [#6016](https://github.com/jj-vcs/jj/issues/6016)
 
+* Rename the `core.fsmonitor` and `core.watchman` settings to
+  `fsmonitor.backend`, and `fsmonitor.watchman` respectively.
+
 ### New features
 
 * `jj workspace list` now accepts `-T`/`--template` option to customize its output via templates.